Started simplifying using constraints in custom plots
Moved the loops over pivots and constraints to an abstract class for lineplot. Now a specific static (matplotlib) plot can just implement a short plot method for plotting one series on an axis which is called from the parent.
